+++ xarray/core/formatting.py
@@ -200,7 +200,22 @@ def _summarize_var_or_coord(name, var, col_width, show_values=True,
         values_str = format_array_flat(var, max_width - len(front_str))
     else:
         values_str = u'...'
+
     return front_str + values_str
+
+
+def _summarize_coord_multiindex(coord, col_width, marker):
+    first_col = pretty_print(u'  %s %s ' % (marker, coord.name), col_width)
+    return u'%s(%s) MultiIndex' % (first_col, unicode_type(coord.dims[0]))
+
+
+def _summarize_coord_levels(coord, col_width, marker=u'-'):
+    relevant_coord = coord[:30]
+    return u'\n'.join(
+        [_summarize_var_or_coord(lname,
+                                 relevant_coord.get_level_variable(lname),
+                                 col_width, marker=marker)
+         for lname in coord.level_names])
 
 
 def _not_remote(var):
@@ -222,6 +237,12 @@ def summarize_coord(name, var, col_width):
     is_index = name in var.dims
     show_values = is_index or _not_remote(var)
     marker = u'*' if is_index else u' '
+    if is_index:
+        coord = var.variable.to_index_variable()
+        if coord.level_names is not None:
+            return u'\n'.join(
+                [_summarize_coord_multiindex(coord, col_width, marker),
+                 _summarize_coord_levels(coord, col_width)])
     return _summarize_var_or_coord(name, var, col_width, show_values, marker)
 
 
@@ -233,9 +254,26 @@ def summarize_attr(key, value, col_width=None):
 EMPTY_REPR = u'    *empty*'
 
 
-def _calculate_col_width(mapping):
-    max_name_length = (max(len(unicode_type(k)) for k in mapping)
-                       if mapping else 0)
+def _get_col_items(mapping):
+    """Get all column items to format, including both keys of `mapping`
+    and MultiIndex levels if any.
+    """
+    from .variable import IndexVariable
+
+    col_items = []
+    for k, v in mapping.items():
+        col_items.append(k)
+        var = getattr(v, 'variable', v)
+        if isinstance(var, IndexVariable):
+            level_names = var.to_index_variable().level_names
+            if level_names is not None:
+                col_items += list(level_names)
+    return col_items
+
+
+def _calculate_col_width(col_items):
+    max_name_length = (max(len(unicode_type(s)) for s in col_items)
+                       if col_items else 0)
     col_width = max(max_name_length, 7) + 6
     return col_width
 
@@ -251,16 +289,19 @@ def _mapping_repr(mapping, title, summarizer, col_width=None):
     return u'\n'.join(summary)
 
 
-coords_repr = functools.partial(_mapping_repr, title=u'Coordinates',
-                                summarizer=summarize_coord)
-
-
 vars_repr = functools.partial(_mapping_repr, title=u'Data variables',
                               summarizer=summarize_var)
 
 
 attrs_repr = functools.partial(_mapping_repr, title=u'Attributes',
                                summarizer=summarize_attr)
+
+
+def coords_repr(coords, col_width=None):
+    if col_width is None:
+        col_width = _calculate_col_width(_get_col_items(coords))
+    return _mapping_repr(coords, title=u'Coordinates',
+                         summarizer=summarize_coord, col_width=col_width)
 
 
 def indexes_repr(indexes):
@@ -302,7 +343,7 @@ def array_repr(arr):
 def dataset_repr(ds):
     summary = [u'<xarray.%s>' % type(ds).__name__]
 
-    col_width = _calculate_col_width(ds)
+    col_width = _calculate_col_width(_get_col_items(ds))
 
     dims_start = pretty_print(u'Dimensions:', col_width)
     all_dim_strings = [u'%s: %s' % (k, v) for k, v in iteritems(ds.dims)]